chore(keymaps): make keymaps load modularly
now keymaps can be defined in config/keymaps
within lua files that return a function that accepts a
map argument which will be shortcut to vim.map.set

Fixes N/A
