+++ lazyvim.json
@@ -3,8 +3,9 @@
     "lazyvim.plugins.extras.lang.markdown",
     "lazyvim.plugins.extras.test.core"
   ],
+  "install_version": 7,
   "news": {
     "NEWS.md": "10960"
   },
-  "version": 7
+  "version": 8
 }
