+++ lua/config/keymaps/luasnip-keys.lua
@@ -1,0 +1,16 @@
+return function(map)
+  -- LuaSnip maps
+
+  vim.api.nvim_set_keymap(
+    "i",
+    "<S-CR>",
+    [[<cmd>lua require('luasnip').expand_or_jump()<CR>]],
+    { noremap = true, silent = true }
+  )
+  vim.api.nvim_set_keymap(
+    "s",
+    "<S-CR>",
+    [[<cmd>lua require('luasnip').expand_or_jump()<CR>]],
+    { noremap = true, silent = true }
+  )
+end
