+++ lua/config/keymaps/multiverse-keys.lua
@@ -1,0 +1,8 @@
+return function(map)
+
+  -- Open Projects Window
+  map("n", "<Space>p", function()
+    vim.cmd(":Telescope workspaces")
+  end)
+
+end
