+++ lua/config/keymaps/test-keys.lua
@@ -1,0 +1,13 @@
+return function(map)
+  map("n", "<leader>tp", function()
+    vim.cmd("w | PlenaryBustedFile %")
+  end)
+
+  map("n", "<leader>tn", function()
+    vim.cmd("w | Neotest run")
+  end)
+
+  map("n", "<leader>ts", function()
+    vim.cmd("w | Neotest summary")
+  end)
+end
